% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp20098.R
\name{getAcc}
\alias{getAcc}
\title{Calculate the prediction accuracy of GMM}
\usage{
getAcc(test_data, k, weight, mu, sigma)
}
\arguments{
\item{test_data}{test set(matrix)}

\item{k}{The number of Gaussian components (numeric)}

\item{weight}{GMM weights on training set (matrix)}

\item{mu}{GMM mu on training set (matrix)}

\item{sigma}{GMM sigma on training set (matrix)}
}
\value{
prediction accuracy on test set (numeric)
}
\description{
Calculate the prediction accuracy of GMM
}
\examples{
\dontrun{
acc <- getAcc(test,2,weight,mu,sigma)
}
}
