% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp20098.R
\name{getInitGMM}
\alias{getInitGMM}
\title{Initializing GMM with k-means}
\usage{
getInitGMM(dat, k)
}
\arguments{
\item{dat}{training set(matrix)}

\item{k}{The number of Gaussian components (numeric)}
}
\value{
a list (weight,mu,sigma)
}
\description{
Initializing GMM with k-means
}
\examples{
\dontrun{
init <- getInitGMM(A_train,2)
}
}
